Issue 207
Additional Font3D and Font2D improvements from Gamalocus (Oak).


git-svn-id: http://jmonkeyengine.googlecode.com/svn/trunk@3037 75d07b2b-3a1a-0410-a2c5-0572b91ccdca

+++ src/com/jmex/font2d/Font2D.java
@@ -60,7 +60,7 @@ public class Font2D implements TextFactory {
         fontTextureState = getFontTextureState(this.fontBitmapFile);
     }
 
-    public Text2D createText(String text, int size, int flags) {
+    public Text2D createText(String text, float size, int flags) {
         Text2D textObj = new Text2D(this, text, size, flags);
         textObj.setCullMode(SceneElement.CULL_NEVER);
         textObj.setRenderState(fontTextureState);
